+++ libyul/optimiser/StackToMemoryMover.cpp
@@ -54,25 +54,22 @@ vector<Statement> generateMemoryStore(
 void StackToMemoryMover::run(
 	OptimiserStepContext& _context,
 	u256 _reservedMemory,
-	map<YulString, map<YulString, uint64_t>> const& _memorySlots,
+	map<YulString, uint64_t> const& _memorySlots,
 	uint64_t _numRequiredSlots,
 	Block& _block
 )
 {
-	StackToMemoryMover stackToMemoryMover(_context, _reservedMemory, _memorySlots, _numRequiredSlots);
+	VariableMemoryOffsetTracker memoryOffsetTracker(_reservedMemory, _memorySlots, _numRequiredSlots);
+	StackToMemoryMover stackToMemoryMover(_context, memoryOffsetTracker);
 	stackToMemoryMover(_block);
 }
 
 StackToMemoryMover::StackToMemoryMover(
 	OptimiserStepContext& _context,
-	u256 _reservedMemory,
-	map<YulString, map<YulString, uint64_t>> const& _memorySlots,
-	uint64_t _numRequiredSlots
+	VariableMemoryOffsetTracker const& _memoryOffsetTracker
 ):
 m_context(_context),
-m_reservedMemory(std::move(_reservedMemory)),
-m_memorySlots(_memorySlots),
-m_numRequiredSlots(_numRequiredSlots),
+m_memoryOffsetTracker(_memoryOffsetTracker),
 m_nameDispenser(_context.dispenser)
 {
 	auto const* evmDialect = dynamic_cast<EVMDialect const*>(&_context.dialect);
@@ -80,41 +77,25 @@ m_nameDispenser(_context.dispenser)
 		evmDialect && evmDialect->providesObjectAccess(),
 		"StackToMemoryMover can only be run on objects using the EVMDialect with object access."
 	);
-
-	if (m_memorySlots.count(YulString{}))
-		// If the global scope contains variables to be moved, start with those as if it were a function.
-		m_currentFunctionMemorySlots = &m_memorySlots.at(YulString{});
 }
 
 void StackToMemoryMover::operator()(FunctionDefinition& _functionDefinition)
 {
-	if (m_memorySlots.count(_functionDefinition.name))
-	{
-		map<YulString, uint64_t> const* saved = m_currentFunctionMemorySlots;
-		m_currentFunctionMemorySlots = &m_memorySlots.at(_functionDefinition.name);
-		for (TypedName const& param: _functionDefinition.parameters + _functionDefinition.returnVariables)
-			if (m_currentFunctionMemorySlots->count(param.name))
-			{
-				// TODO: we cannot handle function parameters yet.
-				m_currentFunctionMemorySlots = saved;
-				return;
-			}
-		ASTModifier::operator()(_functionDefinition);
-		m_currentFunctionMemorySlots = saved;
-	}
+	for (TypedName const& param: _functionDefinition.parameters + _functionDefinition.returnVariables)
+		if (m_memoryOffsetTracker(param.name))
+		{
+			// TODO: we cannot handle function parameters yet.
+			return;
+		}
+	ASTModifier::operator()(_functionDefinition);
 }
 
 void StackToMemoryMover::operator()(Block& _block)
 {
 	using OptionalStatements = std::optional<vector<Statement>>;
-	if (!m_currentFunctionMemorySlots)
-	{
-		ASTModifier::operator()(_block);
-		return;
-	}
 	auto containsVariableNeedingEscalation = [&](auto const& _variables) {
 		return util::contains_if(_variables, [&](auto const& var) {
-			return m_currentFunctionMemorySlots->count(var.name);
+			return m_memoryOffsetTracker(var.name);
 		});
 	};
 	auto rewriteAssignmentOrVariableDeclaration = [&](
@@ -123,12 +104,16 @@ void StackToMemoryMover::operator()(Block& _block)
 		std::unique_ptr<Expression> _value
 	) -> std::vector<Statement> {
 		if (_variables.size() == 1)
+		{
+			optional<YulString> offset = m_memoryOffsetTracker(_variables.front().name);
+			yulAssert(offset, "");
 			return generateMemoryStore(
 				m_context.dialect,
 				_loc,
-				memoryOffset(_variables.front().name),
+				*offset,
 				_value ? *std::move(_value) : Literal{_loc, LiteralKind::Number, "0"_yulstring, {}}
 			);
+		}
 
 		VariableDeclaration tempDecl{_loc, {}, std::move(_value)};
 		vector<Statement> memoryAssignments;
@@ -138,11 +123,11 @@ void StackToMemoryMover::operator()(Block& _block)
 			YulString tempVarName = m_nameDispenser.newName(var.name);
 			tempDecl.variables.emplace_back(TypedName{var.location, tempVarName, {}});
 
-			if (m_currentFunctionMemorySlots->count(var.name))
+			if (optional<YulString> offset = m_memoryOffsetTracker(var.name))
 				memoryAssignments += generateMemoryStore(
 					m_context.dialect,
 					_loc,
-					memoryOffset(var.name),
+					*offset,
 					Identifier{_loc, tempVarName}
 				);
 			else if constexpr (std::is_same_v<std::decay_t<decltype(var)>, Identifier>)
@@ -201,35 +186,36 @@ void StackToMemoryMover::operator()(Block& _block)
 
 void StackToMemoryMover::visit(Expression& _expression)
 {
-	if (
-		Identifier* identifier = std::get_if<Identifier>(&_expression);
-		identifier && m_currentFunctionMemorySlots && m_currentFunctionMemorySlots->count(identifier->name)
-	)
-	{
-		BuiltinFunction const* memoryLoadFunction = m_context.dialect.memoryLoadFunction(m_context.dialect.defaultType);
-		yulAssert(memoryLoadFunction, "");
-		langutil::SourceLocation loc = identifier->location;
-		_expression = FunctionCall{
-			loc,
-			Identifier{loc, memoryLoadFunction->name}, {
-				Literal{
-					loc,
-					LiteralKind::Number,
-					memoryOffset(identifier->name),
-					{}
+	if (Identifier* identifier = std::get_if<Identifier>(&_expression))
+		if (optional<YulString> offset = m_memoryOffsetTracker(identifier->name))
+		{
+			BuiltinFunction const* memoryLoadFunction = m_context.dialect.memoryLoadFunction(m_context.dialect.defaultType);
+			yulAssert(memoryLoadFunction, "");
+			langutil::SourceLocation loc = identifier->location;
+			_expression = FunctionCall{
+				loc,
+				Identifier{loc, memoryLoadFunction->name}, {
+					Literal{
+						loc,
+						LiteralKind::Number,
+						*offset,
+						{}
+					}
 				}
-			}
-		};
+			};
+			return;
+		}
+	ASTModifier::visit(_expression);
+}
+
+optional<YulString> StackToMemoryMover::VariableMemoryOffsetTracker::operator()(YulString _variable) const
+{
+	if (m_memorySlots.count(_variable))
+	{
+		uint64_t slot = m_memorySlots.at(_variable);
+		yulAssert(slot < m_numRequiredSlots, "");
+		return YulString{util::toCompactHexWithPrefix(m_reservedMemory + 32 * (m_numRequiredSlots - slot - 1))};
 	}
 	else
-		ASTModifier::visit(_expression);
+		return std::nullopt;
 }
-
-YulString StackToMemoryMover::memoryOffset(YulString _variable)
-{
-	yulAssert(m_currentFunctionMemorySlots, "");
-	uint64_t slot = m_currentFunctionMemorySlots->at(_variable);
-	yulAssert(slot < m_numRequiredSlots, "");
-	return YulString{util::toCompactHexWithPrefix(m_reservedMemory + 32 * (m_numRequiredSlots - slot - 1))};
-}
-