+++ test/cmdlineTests/yul_unimplemented/input.sol
@@ -1,11 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.0;
 contract test {
-    struct S {
-        uint x;
-    }
-    S str;
-    constructor() {
-        delete str;
+    bytes a;
+    function f() public {
+        delete a;
     }
 }