+++ test/libsolidity/semanticTests/inlineAssembly/inline_assembly_storage_access_via_pointer.sol
@@ -18,6 +18,8 @@ contract C {
     }
 }
 
+// ====
+// compileViaYul: also
 // ----
 // f() -> true
 // a() -> 7