+++ test/libsolidity/syntaxTests/constants/file_level_memory_inverted.sol
@@ -1,0 +1,3 @@
+uint[] constant memory x = 2;
+// ----
+// ParserError 2314: (16-22): Expected identifier but got 'memory'