+++ test/libsolidity/syntaxTests/constants/mapping_constant.sol
@@ -1,0 +1,3 @@
+mapping(uint => uint) constant b = b;
+// ----
+// TypeError 9259: (0-36): Constants of non-value type not yet implemented.