+++ test/libsolidity/syntaxTests/types/mapping/library_storage_parameter_with_nested_mapping.sol
@@ -1,0 +1,5 @@
+library Test {
+    struct Nested { mapping(uint => uint)[2][] a; }
+    struct X { Nested n; }
+    function f(X storage x) public {}
+}