+++ Changelog.md
@@ -7,6 +7,16 @@ Breaking Changes:
 
 ### 0.7.4 (unreleased)
 
+Language Features:
+ * Constants can be defined at file level.
+
+
+Compiler Features:
+
+
+Bugfixes:
+ * Type Checker: Fix internal compiler error caused by storage parameters with nested mappings in libraries.
+
 
 ### 0.7.3 (2020-10-07)